move file uploader agent creation code up to a common location, use the uploader agent for reloadjs as well
