--- src/compiler/custom_js_modules/zlib.ts
@@ -1,4 +1,0 @@
-// TODO this is just a placeholder module so that bundling won't break
-// TODO eventually we would want a real implementation of this in Azle or wasmedge-quickjs or Waden
-
-export {};
